# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog Service API",
        "description": "Service Katalog Buku - Tugas 2 IAE",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "operationId": "ab6891af84b2f0a68a33cdfe8324ce98",
                "responses": {
                    "200": {
                        "description": "Books retrieved successfully"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create new book",
                "operationId": "a1d219764293fcc881a84c059760d6f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "stock"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Harry Potter"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "J.K Rowling"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get book by ID",
                "operationId": "df07027f12397ee981e78f8db169fffa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID buku",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book retrieved successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}